+++ src/lib.rs
@@ -20,8 +20,6 @@
 #![allow(clippy::redundant_closure)]
 
 pub mod config;
-pub mod context;
-pub mod control_plane;
 pub mod error;
 pub mod handlers;
 pub mod policy;